Update problem description and method signature for June 22
This commit changes the name of June 22nd problem to "1248. Count Number of Nice Subarrays" and updates its description accordingly. It also modifies the code by renaming the problem methods as numberOfSubarrays1 and numberOfSubarrays2 and updating their signatures to match the problem's inputs. Furthermore, some clarifications were added to the markdown tutorial document containing explanations and pseudocode for the previous problem.