Implement numberOfSubarrays1 function
The code has been updated to include the full implementation of the numberOfSubarrays1 function. This function finds the total number of nice subarrays in a given list with a specified number of odds. Test cases have also been added for the newly implemented function. A minor formatting update was made to the sliding_window_data append line.