Add detailed logging to 'numberOfSubarrays1' function
The primary change in this commit is the addition of detailed logging to the 'numberOfSubarrays1' function. Enhanced logging was implemented at various stages like input parameters, initialization, counting of even numbers, and the calculation of 'nice subarrays'. This helps in better understanding and debugging of the code. Also, one of the test cases calling this function has been commented out.