Implement minKBitFlips functions
The body of the functions minKBitFlips1, minKBitFlips2, and minKBitFlips3 have been implemented. The implementations are different versions of the same problem, counting the minimum 'k' bit flips required in an array. The functions now check the current state and if required, flip a 'k'-sized window and count the total flips. If the operation cannot be performed, functions return -1.