Update TreeVisualizer class to use custom file names
The TreeVisualizer class has been updated to expect a file name string as an argument in the visualize() method. This provides more control over the output images, allowing them to be named according to the specific functions or conditions that generated them. This change required updating the corresponding function calls in the DailyQuestions script accordingly.