Update problem for June 27, 2024 and simplify functions
The problem for June 27, 2024 is updated from an undefined problem to "1791. Find Center of Star Graph". Corresponding details including example inputs, outputs, and constraints are added. In addition, unnecessary function 'problem4_2' is removed and 'problem4_1' is replaced by a simpler method 'findCenter1'. The test case for the new problem is included in the test section.