Update problem for June 29, 2024, with a new algorithm task
The problem for June 29, 2024, is updated to "2192. All Ancestors of a Node in a Directed Acyclic Graph." The goal is to return a sorted list of ancestors for each node in a Directed Acyclic Graph provided as an array of edges. The problem statement, examples, and constraints have been added to the markdown file and python test cases and function templates are also updated.