Implement `splitSBT3` function in BST problem
This commit introduces the implementation of `splitSBT2` function which splits the BST into two subtrees such
that the first subtree contains nodes with values less than or equal to `target`, and the second subtree contains
nodes with values greater than `target`. It solves it using an iterative approach with two dummies.