Update problem statements and solutions for July
The problem statements and solutions for the first week of July 2024 have been updated in `DailyQuestions/2024/7. Jul/Jul01-07.py`. The problem for July 1st, "1550. Three Consecutive Odds", has been added with its solution. The placeholders for the problem statements and solutions for the rest of the week have been updated with the correct dates.